refactor(storage): extract content/skills/history modules (R.4)
Phase R.4 of the storage modularization plan
(SPEC_STORE_MODULARIZATION_2026_05_27.md). Three sibling subsystems
extracted in one PR — each is small (~100-150 LOC) and self-
contained:

- `content.rs` (111 LOC): AgentContent struct + 4 agent_content_*
  methods.
- `skills.rs` (133 LOC): AgentSkill struct + 5 agent_skill_*
  methods.
- `history.rs` (160 LOC): AgentHistory struct + 3 agent_history_*
  methods, plus the `format_epoch_date` helper that's only used by
  `agent_history_append`.

Method surface is unchanged: `Store::agent_content_*`,
`agent_skill_*`, `agent_history_*` still live on `Store` thanks to
sibling `impl Store {}` blocks. The 3 structs are re-exported from
`storage::store` AND from `storage::` so existing imports keep
working without churn at call sites
(e.g. `use crate::backend::storage::store::AgentSkill` in
`agent_config.rs`).

Net store.rs change: -323 lines. New files: 404 LOC combined.

162/162 storage tests green.
