+++ .changesets/1779972519-refactor-storage-extract-content-skills-history-modules-r-4-8zg4.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(storage): extract content/skills/history modules (R.4)
