+++ agentmux-srv/src/backend/storage/store.rs
@@ -575,36 +575,13 @@ fn default_agent_type() -> String {
 }
 
 /// A content blob associated with a agent definition.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentContent {
-    pub agent_id: String,
-    pub content_type: String,
-    pub content: String,
-    pub updated_at: i64,
-}
-
-/// A reusable skill/capability attached to a agent definition.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentSkill {
-    pub id: String,
-    pub agent_id: String,
-    pub name: String,
-    pub trigger: String,
-    pub skill_type: String,
-    pub description: String,
-    pub content: String,
-    pub created_at: i64,
-}
-
-/// An append-only session history entry for a agent definition.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AgentHistory {
-    pub id: i64,
-    pub agent_id: String,
-    pub session_date: String,
-    pub entry: String,
-    pub timestamp: i64,
-}
+// AgentContent / AgentSkill / AgentHistory moved to
+// `super::content` / `super::skills` / `super::history` in Phase R.4
+// (SPEC_STORE_MODULARIZATION_2026_05_27.md). Re-exported here so
+// external `storage::store::AgentContent` etc. imports keep working.
+pub use super::content::AgentContent;
+pub use super::history::AgentHistory;
+pub use super::skills::AgentSkill;
 
 impl Store {
     /// List all agent definitions, **most-recently-used first**.
@@ -1023,306 +1000,11 @@ impl Store {
         Ok(rows > 0)
     }
 
-    // ---- AgentContent CRUD ----
-
-    /// Get a single content blob for an agent.
-    pub fn agent_content_get(&self, agent_id: &str, content_type: &str) -> Result<Option<AgentContent>, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn.prepare(
-            "SELECT agent_id, content_type, content, updated_at
-             FROM db_agent_content WHERE agent_id=?1 AND content_type=?2",
-        )?;
-        let result = stmt.query_row(params![agent_id, content_type], |row| {
-            Ok(AgentContent {
-                agent_id: row.get(0)?,
-                content_type: row.get(1)?,
-                content: row.get(2)?,
-                updated_at: row.get(3)?,
-            })
-        });
-        match result {
-            Ok(content) => Ok(Some(content)),
-            Err(rusqlite::Error::QueryReturnedNoRows) => Ok(None),
-            Err(e) => Err(StoreError::Sqlite(e)),
-        }
-    }
-
-    /// Upsert a content blob for an agent.
-    pub fn agent_content_set(&self, content: &AgentContent) -> Result<(), StoreError> {
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "INSERT INTO db_agent_content (agent_id, content_type, content, updated_at)
-             VALUES (?1, ?2, ?3, ?4)
-             ON CONFLICT(agent_id, content_type) DO UPDATE SET content=?3, updated_at=?4",
-            params![
-                content.agent_id,
-                content.content_type,
-                content.content,
-                content.updated_at,
-            ],
-        )?;
-        Ok(())
-    }
-
-    /// Get all content blobs for an agent.
-    pub fn agent_content_get_all(&self, agent_id: &str) -> Result<Vec<AgentContent>, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn.prepare(
-            "SELECT agent_id, content_type, content, updated_at
-             FROM db_agent_content WHERE agent_id=?1 ORDER BY content_type ASC",
-        )?;
-        let rows = stmt.query_map(params![agent_id], |row| {
-            Ok(AgentContent {
-                agent_id: row.get(0)?,
-                content_type: row.get(1)?,
-                content: row.get(2)?,
-                updated_at: row.get(3)?,
-            })
-        })?;
-        let mut contents = Vec::new();
-        for row in rows {
-            contents.push(row?);
-        }
-        Ok(contents)
-    }
-
-    /// Delete a specific content blob. Returns true if a row was deleted.
-    #[allow(dead_code)]
-    pub fn agent_content_delete(&self, agent_id: &str, content_type: &str) -> Result<bool, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let rows = conn.execute(
-            "DELETE FROM db_agent_content WHERE agent_id=?1 AND content_type=?2",
-            params![agent_id, content_type],
-        )?;
-        Ok(rows > 0)
-    }
-
-    // ---- AgentSkill CRUD ----
-
-    /// List all skills for an agent, ordered by created_at ascending.
-    pub fn agent_skill_list(&self, agent_id: &str) -> Result<Vec<AgentSkill>, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn.prepare(
-            "SELECT id, agent_id, name, trigger, skill_type, description, content, created_at
-             FROM db_agent_skills WHERE agent_id=?1 ORDER BY created_at ASC",
-        )?;
-        let rows = stmt.query_map(params![agent_id], |row| {
-            Ok(AgentSkill {
-                id: row.get(0)?,
-                agent_id: row.get(1)?,
-                name: row.get(2)?,
-                trigger: row.get(3)?,
-                skill_type: row.get(4)?,
-                description: row.get(5)?,
-                content: row.get(6)?,
-                created_at: row.get(7)?,
-            })
-        })?;
-        let mut skills = Vec::new();
-        for row in rows {
-            skills.push(row?);
-        }
-        Ok(skills)
-    }
-
-    /// Get a single skill by id.
-    pub fn agent_skill_get(&self, id: &str) -> Result<Option<AgentSkill>, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let mut stmt = conn.prepare(
-            "SELECT id, agent_id, name, trigger, skill_type, description, content, created_at
-             FROM db_agent_skills WHERE id=?1",
-        )?;
-        let result = stmt.query_row(params![id], |row| {
-            Ok(AgentSkill {
-                id: row.get(0)?,
-                agent_id: row.get(1)?,
-                name: row.get(2)?,
-                trigger: row.get(3)?,
-                skill_type: row.get(4)?,
-                description: row.get(5)?,
-                content: row.get(6)?,
-                created_at: row.get(7)?,
-            })
-        });
-        match result {
-            Ok(skill) => Ok(Some(skill)),
-            Err(rusqlite::Error::QueryReturnedNoRows) => Ok(None),
-            Err(e) => Err(StoreError::Sqlite(e)),
-        }
-    }
-
-    /// Insert a new skill.
-    pub fn agent_skill_insert(&self, skill: &AgentSkill) -> Result<(), StoreError> {
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "INSERT INTO db_agent_skills (id, agent_id, name, trigger, skill_type, description, content, created_at)
-             VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8)",
-            params![
-                skill.id,
-                skill.agent_id,
-                skill.name,
-                skill.trigger,
-                skill.skill_type,
-                skill.description,
-                skill.content,
-                skill.created_at
-            ],
-        )?;
-        Ok(())
-    }
-
-    /// Update an existing skill (all fields except id, agent_id, created_at).
-    pub fn agent_skill_update(&self, skill: &AgentSkill) -> Result<bool, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let rows = conn.execute(
-            "UPDATE db_agent_skills SET name=?1, trigger=?2, skill_type=?3, description=?4, content=?5
-             WHERE id=?6",
-            params![
-                skill.name,
-                skill.trigger,
-                skill.skill_type,
-                skill.description,
-                skill.content,
-                skill.id
-            ],
-        )?;
-        Ok(rows > 0)
-    }
-
-    /// Delete a skill by id. Returns true if a row was deleted.
-    pub fn agent_skill_delete(&self, id: &str) -> Result<bool, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let rows = conn.execute(
-            "DELETE FROM db_agent_skills WHERE id=?1",
-            params![id],
-        )?;
-        Ok(rows > 0)
-    }
-
-    // ---- AgentHistory methods ----
-
-    /// Append a history entry for an agent. Auto-sets session_date (today) and timestamp.
-    pub fn agent_history_append(&self, agent_id: &str, entry: &str) -> Result<AgentHistory, StoreError> {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_millis() as i64;
-        // session_date as YYYY-MM-DD
-        let secs = (now / 1000) as u64;
-        let days = secs / 86400;
-        // Simple date calculation (no chrono dependency needed)
-        let session_date = format_epoch_date(days);
-        let conn = self.conn.lock().unwrap();
-        conn.execute(
-            "INSERT INTO db_agent_history (agent_id, session_date, entry, timestamp) VALUES (?1, ?2, ?3, ?4)",
-            params![agent_id, session_date, entry, now],
-        )?;
-        let id = conn.last_insert_rowid();
-        Ok(AgentHistory {
-            id,
-            agent_id: agent_id.to_string(),
-            session_date,
-            entry: entry.to_string(),
-            timestamp: now,
-        })
-    }
-
-    /// List history entries for an agent, with optional date filter and pagination.
-    pub fn agent_history_list(
-        &self,
-        agent_id: &str,
-        session_date: Option<&str>,
-        limit: i64,
-        offset: i64,
-    ) -> Result<Vec<AgentHistory>, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let (sql, params_vec): (String, Vec<Box<dyn rusqlite::types::ToSql>>) = match session_date {
-            Some(date) => (
-                "SELECT id, agent_id, session_date, entry, timestamp
-                 FROM db_agent_history WHERE agent_id=?1 AND session_date=?2
-                 ORDER BY timestamp DESC LIMIT ?3 OFFSET ?4".to_string(),
-                vec![
-                    Box::new(agent_id.to_string()),
-                    Box::new(date.to_string()),
-                    Box::new(limit),
-                    Box::new(offset),
-                ],
-            ),
-            None => (
-                "SELECT id, agent_id, session_date, entry, timestamp
-                 FROM db_agent_history WHERE agent_id=?1
-                 ORDER BY timestamp DESC LIMIT ?2 OFFSET ?3".to_string(),
-                vec![
-                    Box::new(agent_id.to_string()),
-                    Box::new(limit),
-                    Box::new(offset),
-                ],
-            ),
-        };
-        let params_refs: Vec<&dyn rusqlite::types::ToSql> = params_vec.iter().map(|b| b.as_ref()).collect();
-        let mut stmt = conn.prepare(&sql)?;
-        let rows = stmt.query_map(params_refs.as_slice(), |row| {
-            Ok(AgentHistory {
-                id: row.get(0)?,
-                agent_id: row.get(1)?,
-                session_date: row.get(2)?,
-                entry: row.get(3)?,
-                timestamp: row.get(4)?,
-            })
-        })?;
-        let mut entries = Vec::new();
-        for row in rows {
-            entries.push(row?);
-        }
-        Ok(entries)
-    }
-
-    /// Search history entries for an agent using LIKE-based matching.
-    pub fn agent_history_search(
-        &self,
-        agent_id: &str,
-        query: &str,
-        limit: i64,
-    ) -> Result<Vec<AgentHistory>, StoreError> {
-        let conn = self.conn.lock().unwrap();
-        let pattern = format!("%{}%", query);
-        let mut stmt = conn.prepare(
-            "SELECT id, agent_id, session_date, entry, timestamp
-             FROM db_agent_history WHERE agent_id=?1 AND entry LIKE ?2
-             ORDER BY timestamp DESC LIMIT ?3",
-        )?;
-        let rows = stmt.query_map(params![agent_id, pattern, limit], |row| {
-            Ok(AgentHistory {
-                id: row.get(0)?,
-                agent_id: row.get(1)?,
-                session_date: row.get(2)?,
-                entry: row.get(3)?,
-                timestamp: row.get(4)?,
-            })
-        })?;
-        let mut entries = Vec::new();
-        for row in rows {
-            entries.push(row?);
-        }
-        Ok(entries)
-    }
-}
-
-/// Format days-since-epoch as YYYY-MM-DD string.
-/// Simple implementation without chrono dependency.
-fn format_epoch_date(days_since_epoch: u64) -> String {
-    // Algorithm from https://howardhinnant.github.io/date_algorithms.html
-    let z = days_since_epoch + 719468;
-    let era = z / 146097;
-    let doe = z - era * 146097;
-    let yoe = (doe - doe / 1460 + doe / 36524 - doe / 146096) / 365;
-    let y = yoe + era * 400;
-    let doy = doe - (365 * yoe + yoe / 4 - yoe / 100);
-    let mp = (5 * doy + 2) / 153;
-    let d = doy - (153 * mp + 2) / 5 + 1;
-    let m = if mp < 10 { mp + 3 } else { mp - 9 };
-    let y = if m <= 2 { y + 1 } else { y };
-    format!("{:04}-{:02}-{:02}", y, m, d)
+    // AgentContent / AgentSkill / AgentHistory CRUD moved to
+    // `super::content` / `super::skills` / `super::history` in Phase R.4
+    // (SPEC_STORE_MODULARIZATION_2026_05_27.md). The `impl Store {}`
+    // blocks there add the same methods to this type. `format_epoch_date`
+    // moves with `agent_history_append`.
 }
 
 // ====================================================================
